Header and navbar are files for easy inclusion
I've made the header portion with all the javascript and css files needed for everything available as the header.php file.  Just include that in your code to use it.  Also, I did the same with the navbar at the top of the screen.  This way, if there are any changes made to the nav bar, I only have to change one instead of every page that has the navigation bar.